changed input structure and added seperate utility files
What changes you have made?
changed input structure from array to object
And json utility and file access utility files from utility library

Why this change is necessary?
for accessing options by their names objects are better.
To remove the complexity of keeping all functions in same file

What is the impact of the change?
easy to access the user inputs.
And easy to use and check the utility functions