added isFileExists function
What changes you have made?
added is file exists in query and save function
calling the read,write and is file exists function from the main

Why this change is necessary?
to make sure functions are working properly

What is the impact of the change?
files can be called from the main function