modified data structure and changed test cases.
What changes you have made?
changed the data structure and the code accordingly
modified the test cases according to the data structure

Why this change is necessary?
to handle the requirements in the better way

What is the impact of the change?
need to change some test cases as the structure is changed
as well as needed more test cases