deleted unwanted functions and loops
What changes you have made?
deleted some unwanted lines in the code
changed some let's to const's

Why this change is necessary?
as the unwanted lines looks messy
and some variables are not changed in the function so declared as const

What is the impact of the change?
now code looks little better than before